Moves spectrum from montecarlo
